feat(api): add OAuth token fetching (#8)
Override `prepareOptions` (since it is the earliest async request hook
we have) to obtain a token before the first API request is made. Set
this as the token for the client so that all other API requests get this
added by `authHeaders`.

Update `validateHeaders` to ignore validation of the token request
headers - we won't have a token when obtaining a token.

Update the README to remove `token` and provide examples of the `apiKey`
and `apiSecret` env variables.